# feat(electric): Proposed speedup of Sent Rows Graph popping method