--- a/.changeset/quick-crabs-accept.md
+++ b/.changeset/quick-crabs-accept.md
@@ -1,0 +1,5 @@
+---
+"@core/electric": patch
+---
+
+Improve performance of shape unsubscribe API with alternative `SentRowsGraph` pruning method.