--- a/components/electric/lib/electric/replication/shapes/sent_rows_graph.ex
+++ b/components/electric/lib/electric/replication/shapes/sent_rows_graph.ex
@@ -81,8 +81,6 @@ defmodule Electric.Replication.Shapes.SentRowsGraph do
   defp do_pop_by_request_id(%Graph{} = graph, %MapSet{} = request_ids, root_vertex) do
     predicate = fn {id, _} -> MapSet.member?(request_ids, id) end
 
-    start = System.monotonic_time()
-
     {edges, vertices} =
       dfs_traverse(
         [Graph.Utils.vertex_id(root_vertex)],
@@ -99,9 +97,9 @@ defmodule Electric.Replication.Shapes.SentRowsGraph do
             end)
             |> Enum.split_with(&predicate.(elem(&1, 2)))
             |> case do
-              {all_edges, []} ->
+              {new_edges, []} ->
                 # If all incoming edges match the request ID, we'll pop the vertex
-                {:next, {all_edges ++ edges, [v | vertices]}}
+                {:next, {new_edges ++ edges, [v | vertices]}}
 
               {new_edges, _rest} ->
                 # If some incoming edges are unaffected, we'll pop the edges explicitly
@@ -111,21 +109,16 @@ defmodule Electric.Replication.Shapes.SentRowsGraph do
         fn meta -> any_key_matches_predicate?(meta, predicate) end
       )
 
-    Logger.warning("Traversing graph took #{System.convert_time_unit(System.monotonic_time() - start, :native, :millisecond)}")
-    start = System.monotonic_time()
 
+    # Remove all edges relating to the request IDs from the graph
     graph =
       edges
       |> Enum.reduce(graph, fn {v1, v2, label}, acc -> Graph.delete_edge(acc, v1, v2, label) end)
 
-    Logger.warning("Removing edges from graph took #{System.convert_time_unit(System.monotonic_time() - start, :native, :millisecond)}")
-    start = System.monotonic_time()
+    # Retain the maximally connected subgraph that does not contain the
+    # vertices that have been popped
     vertices_to_keep = MapSet.difference(MapSet.new(Graph.vertices(graph)), MapSet.new(vertices)) |> MapSet.to_list
     graph = Graph.subgraph(graph, vertices_to_keep)
-    # graph = graph
-    #   |> Graph.delete_vertices(vertices)
-
-    Logger.warning("Removing vertices graph took #{System.convert_time_unit(System.monotonic_time() - start, :native, :millisecond)}")
 
     {vertices, graph}
   end